+++ src/Cwaff-Misc/CwaffSynergies.cs
@@ -131,6 +131,10 @@ public static class CwaffSynergies
         // Card speed is dramatically increased, clip size is doubled, and the last 13 cards in each clip become exploding jokers.
         NewMastery<MasteryOfBlackjack>(MASTERY_BLACKJACK, Blackjack.ItemName)
             .MultSpread(0.5f).MultClipSize(2f).MultFireRate(2f);
+        // English can be charged two launch two additional rows of 6 and 7 balls, respectively.
+        NewMastery<MasteryOfEnglish>(MASTERY_ENGLISH, English.ItemName);
+        // Launches 3 knives at a time for no additional cost.
+        NewMastery<MasteryOfIronMaid>(MASTERY_IRON_MAID, IronMaid.ItemName);
       #endregion
 
         SanityCheckAllSynergiesHaveBeenInitialized();
@@ -359,6 +363,8 @@ internal class MasteryOfPistolWhip      : MasteryDummyItem {}
 internal class MasteryOfFemtobyte       : MasteryDummyItem {}
 internal class MasteryOfUppskeruvel     : MasteryDummyItem {}
 internal class MasteryOfBlackjack       : MasteryDummyItem {}
+internal class MasteryOfEnglish         : MasteryDummyItem {}
+internal class MasteryOfIronMaid        : MasteryDummyItem {}
 
 public enum Synergy {
     // Synergies
@@ -410,4 +416,6 @@ public enum Synergy {
     MASTERY_FEMTOBYTE,
     MASTERY_UPPSKERUVEL,
     MASTERY_BLACKJACK,
+    MASTERY_ENGLISH,
+    MASTERY_IRON_MAID,
 };